Rename Prompt Store Provider
